On a second thought, I revert the last commit that removes 'unused code'. This code was made with a purpose, and the reasoning is this:
Let's stay we use a Seq for the Proyection state, specifically Seq[(YearMonth, OrdersAmount)].

If we do so, then when we are updating the Proyection state, the one we are going to use later during the summarization,
that is as easy as performing a concatenation of Sequences. You take the state from the Writeside pipeline, call the
.toSeq method, and easy enough, you just concatenate that to the state of the Proyection, which is a Seq as well.

Why is this not a good idea?

Well, because then the state of the Proyection state grows over time. Instead of having a Map[YearMonth, OrdersAmount] you have
a Seq, and that is growing, constantly.

Not a good idea, really!

But let's say that you use a Map[YearMonth, OrdersAmount] for the Proyection state.

Then updating it is not so easy, because you may have to take each new entry of the Writeside state and merge it with the
Readside state.

You have to merge it, that is the not-easy part, for that I did use zio.prelude operators such as <> which is the
Association operator. However once you get over how it was not trivial, now you have the performance gains:

If you use a Map for the Proyection state, now the summarization step will have an incredibly finite amount of data
to summarize. It will go over the Map[YearMonth, OrdersAmount], and classify the keys, and the map entries will be
very very few.

This is the way to do it.

Sorry for the rollback, I am adding this code back online.

Best regards